Add switch  append-only and delete-load functionality

+++ operators/load_data.py
@@ -16,17 +16,20 @@ class LoadDataOperator(BaseOperator):
                  table,
                  redshift_conn_id,
                  sql_insert_script,
+                 is_delete_load=False,
                  *args, **kwargs):
         super(LoadDataOperator, self).__init__(*args, **kwargs)
         self.redshift_conn_id = redshift_conn_id
         self.sql_insert_script = sql_insert_script
         self.table = table
+        self.is_delete_load = is_delete_load
 
     def execute(self, context):
         redshift = PostgresHook(postgres_conn_id=self.redshift_conn_id)
 
-        self.log.info(f"Clearing data from destination Redshift {self.table} table")
-        redshift.run("DELETE FROM {}".format(self.table))
+        if self.is_delete_load:
+            self.log.info(f"Clearing data from destination Redshift {self.table} table")
+            redshift.run("DELETE FROM {}".format(self.table))
 
         self.log.info(f"Start load fact data {self.table}")
         formatted_sql = LoadDataOperator.facts_sql_template.format(
